+++ src/app/controller/common.go
@@ -2,10 +2,14 @@ package controller
 
 import (
 	"encoding/json"
+	"io/ioutil"
 	"net/http"
 
 	"github.com/davyzhang/agw"
 )
+
+const GRAPHQL_API = "https://api.github.com/graphql"
+const LABEL_PREVIEW_HEADER = "application/vnd.github.bane-preview+json"
 
 func sendJson(w http.ResponseWriter, status int, payload interface{}) {
 	response, err := json.Marshal(payload)
@@ -15,4 +19,12 @@ func sendJson(w http.ResponseWriter, status int, payload interface{}) {
 	}
 	w.WriteHeader(status)
 	agw.WriteResponse(w, []byte(response), false)
+}
+
+func getSqlFrom(file string) (string, error) {
+	sql, err := ioutil.ReadFile(file)
+	if err != nil {
+		return "", err
+	}
+	return string(sql), nil
 }