+++ src/queries/sql/get_scrumify_labels.sql
@@ -1,0 +1,7 @@
+SELECT
+  id,
+  name,
+  color,
+  description
+FROM
+  ScrumifyLabels;